// model/network.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GraphTracks Analytics API
 * Bluesky Analytics API for GraphTracks
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * The network to get stats for. Currently only BlueSky is supported.
 * @export
 * @enum {string}
 */

export const Network = {
    BlueSky: 'BlueSky'
} as const;

export type Network = typeof Network[keyof typeof Network];