// model/network-account.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GraphTracks Analytics API
 * Bluesky Analytics API for GraphTracks
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface NetworkAccount
 */
export interface NetworkAccount {
    /**
     * The account id (did for BlueSky)
     * @type {string}
     * @memberof NetworkAccount
     */
    'account_id': string;
    /**
     * The account handle
     * @type {string}
     * @memberof NetworkAccount
     */
    'account_handle': string;
    /**
     * The internal id used by GraphTracks
     * @type {number}
     * @memberof NetworkAccount
     */
    'account_internal_id': number;
    /**
     * Whether the account is active
     * @type {boolean}
     * @memberof NetworkAccount
     */
    'active': boolean;
    /**
     * The timestamp for when the account was created
     * @type {string}
     * @memberof NetworkAccount
     */
    'created_at': string;
    /**
     * The timestamp for when the account was last updated
     * @type {string}
     * @memberof NetworkAccount
     */
    'updated_at': string;
    /**
     * The network this account belongs to
     * @type {string}
     * @memberof NetworkAccount
     */
    'network'?: NetworkAccountNetworkEnum;
}

export const NetworkAccountNetworkEnum = {
    BlueSky: 'BlueSky'
} as const;

export type NetworkAccountNetworkEnum = typeof NetworkAccountNetworkEnum[keyof typeof NetworkAccountNetworkEnum];