// model/timeframe.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GraphTracks Analytics API
 * Bluesky Analytics API for GraphTracks
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * The timeframe to get stats for
 * @export
 * @enum {string}
 */

export const Timeframe = {
    _1h: '1h',
    _1d: '1d',
    _7d: '7d',
    _30d: '30d',
    _90d: '90d',
    All: 'all'
} as const;

export type Timeframe = typeof Timeframe[keyof typeof Timeframe];